Grote update
Weet ik veel, lang geleden maar alles werkt nu redelijk.

+++ api/database/migrations/0001_01_01_000000_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('height')->nullable();
             $table->decimal('weight')->nullable();
             $table->integer('age')->nullable();
-            $table->integer('bmi')->nullable();
+            $table->decimal('bmi')->nullable();
             $table->boolean('lock')->default(false);
             $table->boolean('is_admin')->default(false);
             $table->boolean('is_owner')->default(false);
